--- minimark/src/rmrk/constants.ts
@@ -1,2 +1,0 @@
-export const SQUARE = '::'
-export const RMRK = /^[rR][mM][rR][kK]::/
