+++ minimark/src/rmrk/shared/constants.ts
@@ -1,0 +1,5 @@
+export const SQUARE = '::'
+export const RMRK = /^[rR][mM][rR][kK]::/
+export const RMRK_V0 = '0.0.1'
+export const RMRK_V1 = '1.0.0'
+export const RMRK_V2 = '2.0.0'
