+++ minimark/src/rmrk/unwrap.ts
@@ -1,9 +1,16 @@
 import { decodeHex } from '../utils/hex'
 import { unwrapURI } from '../utils/unwrap'
-import { isRemark, resolveValue, splitBySquare, toInteraction, toVersion } from './helpers'
-import { InteractionValue, UnwrappedRemark } from './types'
+import { isRemark, splitBySquare, toVersion } from './shared/helpers'
+import { RMRK_V0, RMRK_V1, RMRK_V2 } from './shared/constants'
 
-const unwrap = <T = InteractionValue>(text: string): UnwrappedRemark<T | InteractionValue> => {
+import { InteractionV2, InteractionValue, UnwrappedRemark } from './types'
+import { unwrapRemark } from './v1/unwrap'
+import { InteractionV2Value, unwrapRemarkV2 } from './v2'
+
+type V1 = InteractionValue
+type V2 = InteractionV2 | InteractionV2Value
+
+const unwrap = <T = InteractionValue>(text: string): UnwrappedRemark<T | V1 | V2 > => {
   const decoded = unwrapURI(decodeHex(text))
 
   if (!isRemark(decoded)) {
@@ -11,21 +18,19 @@ const unwrap = <T = InteractionValue>(text: string): UnwrappedRemark<T | Interac
   }
 
   // DEV: skipping prefix intetionally
-  const [, mayInteraction, mayVersion, mayIdOrValue, mayValue] = splitBySquare(decoded)
+  const [, , mayVersion] = splitBySquare(decoded)
 
-  const interaction = toInteraction(mayInteraction)
+  const version = toVersion(mayVersion)
 
-  const value = resolveValue<T>(interaction, mayIdOrValue, mayValue)
-
-  if (!value) {
-    throw new TypeError(`RMRK: Unable to unwrap value ${decoded}`)
+  if ([RMRK_V0, RMRK_V1].includes(version)) {
+    return unwrapRemark(text)
   }
 
-  return {
-    interaction,
-    value,
-    version: toVersion(mayVersion)
+  if (version === RMRK_V2) {
+    return unwrapRemarkV2(text) as UnwrappedRemark<T | V1 | V2>
   }
+
+  throw new TypeError(`RMRK: Unable to unwrap version ${version}`)
 }
 
 export { unwrap as unwrapRemark }
